+++ src/main/source-control/hosted-review-creation.test.ts
@@ -109,7 +109,7 @@ vi.mock('./hosted-review', () => ({
   getHostedReviewForBranch: getHostedReviewForBranchMock
 }))
 
-import { createHostedReview, getHostedReviewCreationEligibility } from './hosted-review-creation'
+import { createHostedReview } from './hosted-review-creation'
 
 function resetMocks(): void {
   for (const mock of [
@@ -575,306 +575,5 @@ describe('createHostedReview', () => {
       }
     })
     expect(createGitHubPullRequestMock).not.toHaveBeenCalled()
-  })
-})
-
-describe('getHostedReviewCreationEligibility', () => {
-  beforeEach(() => {
-    resetMocks()
-
-    mockGitHubProvider()
-    getHostedReviewForBranchMock.mockResolvedValue(null)
-    ghExecFileAsyncMock.mockResolvedValue({ stdout: '', stderr: '' })
-    gitExecFileAsyncMock.mockResolvedValue({ stdout: 'Feature title\n', stderr: '' })
-    isAzureDevOpsReviewCreationAuthenticatedMock.mockReturnValue(true)
-    isGiteaReviewCreationAuthenticatedMock.mockReturnValue(true)
-  })
-
-  it('treats short remote base refs as the default branch name', async () => {
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'main',
-        base: 'origin/main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      canCreate: false,
-      blockedReason: 'default_branch',
-      defaultBaseRef: 'origin/main'
-    })
-  })
-
-  // Stacked-worktree base resolution (Change 1/2). `stackedArgs` defaults to a
-  // bare local-only parent; `mockRefs` controls the remote-tracking snapshot.
-  const stackedArgs = (
-    overrides: Partial<Parameters<typeof getHostedReviewCreationEligibility>[0]> = {}
-  ): Parameters<typeof getHostedReviewCreationEligibility>[0] => ({
-    repoPath: '/repo',
-    branch: 'feature/stacked',
-    base: 'stacked-parent',
-    hasUncommittedChanges: false,
-    hasUpstream: true,
-    ahead: 0,
-    behind: 0,
-    ...overrides
-  })
-
-  const mockRefs = (opts: {
-    symbolicRef?: string
-    forEachRef?: string
-    forEachThrows?: boolean
-    revParseThrows?: boolean
-  }): void => {
-    gitExecFileAsyncMock.mockImplementation(async (args: string[]) => {
-      if (args[0] === 'symbolic-ref') {
-        return { stdout: opts.symbolicRef ?? '', stderr: '' }
-      }
-      if (args[0] === 'for-each-ref') {
-        if (opts.forEachThrows) {
-          throw new Error('ssh: connect: connection refused')
-        }
-        return { stdout: opts.forEachRef ?? '', stderr: '' }
-      }
-      if (args[0] === 'rev-parse' && opts.revParseThrows) {
-        throw new Error('unknown revision')
-      }
-      return { stdout: 'refs/remotes/origin/main\n', stderr: '' }
-    })
-  }
-
-  it('falls back to the repo default when a stacked parent base is local-only', async () => {
-    mockRefs({ symbolicRef: 'refs/remotes/origin/main\n' })
-    await expect(getHostedReviewCreationEligibility(stackedArgs())).resolves.toMatchObject({
-      canCreate: true,
-      blockedReason: null,
-      defaultBaseRef: 'origin/main'
-    })
-  })
-
-  it('preserves a stacked parent base that exists on the remote', async () => {
-    mockRefs({ forEachRef: 'refs/remotes/origin/parent-pushed\n' })
-    await expect(
-      getHostedReviewCreationEligibility(stackedArgs({ base: 'parent-pushed' }))
-    ).resolves.toMatchObject({
-      canCreate: true,
-      blockedReason: null,
-      defaultBaseRef: 'parent-pushed'
-    })
-  })
-
-  it('keeps the candidate base when no repo default can be resolved', async () => {
-    mockRefs({ revParseThrows: true })
-    await expect(getHostedReviewCreationEligibility(stackedArgs())).resolves.toMatchObject({
-      canCreate: true,
-      blockedReason: null,
-      defaultBaseRef: 'stacked-parent'
-    })
-  })
-
-  it('preserves the candidate base when the remote probe cannot reach the host', async () => {
-    // Transport failure must not be read as "absent" — that would demote a
-    // legitimately-pushed parent to the repo default on a transient SSH blip.
-    mockRefs({ forEachThrows: true })
-    await expect(
-      getHostedReviewCreationEligibility(stackedArgs({ base: 'parent-pushed' }))
-    ).resolves.toMatchObject({ canCreate: true, defaultBaseRef: 'parent-pushed' })
-  })
-
-  it('blocks dirty tracked GitHub branches before PR creation', async () => {
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'feature/create-pr',
-        base: 'main',
-        hasUncommittedChanges: true,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'github',
-      canCreate: false,
-      blockedReason: 'dirty',
-      nextAction: 'commit',
-      head: 'feature/create-pr'
-    })
-  })
-
-  it('keeps dirty feature branches eligible for PR preparation when review lookup fails', async () => {
-    getHostedReviewForBranchMock.mockRejectedValueOnce(new Error('gh lookup failed'))
-
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'feature/create-pr',
-        base: 'main',
-        hasUncommittedChanges: true,
-        hasUpstream: false,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'github',
-      canCreate: false,
-      blockedReason: 'dirty',
-      nextAction: 'commit',
-      head: 'feature/create-pr'
-    })
-  })
-
-  it('enables creation for clean, in-sync, authenticated GitHub feature branches', async () => {
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'refs/heads/feature/create-pr',
-        base: 'origin/main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'github',
-      canCreate: true,
-      blockedReason: null,
-      nextAction: null,
-      defaultBaseRef: 'origin/main',
-      head: 'feature/create-pr'
-    })
-  })
-
-  it('resolves remote eligibility through SSH repo metadata without generating PR copy', async () => {
-    const remoteGit = {
-      exec: vi.fn(async () => ({ stdout: '', stderr: '' }))
-    }
-    getSshGitProviderMock.mockReturnValue(remoteGit)
-
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/remote/repo',
-        connectionId: 'ssh-1',
-        branch: 'feature/create-pr',
-        base: 'origin/main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'github',
-      canCreate: true,
-      head: 'feature/create-pr'
-    })
-
-    expect(getProjectSlugMock).toHaveBeenCalledWith('/remote/repo', 'ssh-1')
-    expect(getRepoSlugMock).toHaveBeenCalledWith('/remote/repo', 'ssh-1')
-    expect(getHostedReviewForBranchMock).toHaveBeenCalledWith(
-      expect.objectContaining({ repoPath: '/remote/repo', connectionId: 'ssh-1' })
-    )
-    // Why: the base-on-remote probe must run on the SSH host that will execute
-    // the provider create, so it flows through the relay exec, not local git.
-    expect(remoteGit.exec).toHaveBeenCalledWith(
-      ['for-each-ref', '--count=1', '--format=%(refname)', 'refs/remotes/*/main'],
-      '/remote/repo'
-    )
-  })
-
-  it('offers push as the next action for authenticated branches with local-only commits', async () => {
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'feature/create-pr',
-        base: 'main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 2,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      canCreate: false,
-      blockedReason: 'needs_push',
-      nextAction: 'push'
-    })
-  })
-
-  it('enables creation for clean, in-sync, authenticated GitLab feature branches', async () => {
-    mockGitLabProvider()
-
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'feature/gitlab',
-        base: 'main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'gitlab',
-      canCreate: true,
-      blockedReason: null,
-      nextAction: null,
-      head: 'feature/gitlab'
-    })
-    expect(ghExecFileAsyncMock).not.toHaveBeenCalled()
-    expect(glabExecFileAsyncMock).toHaveBeenCalledWith(
-      ['auth', 'status', '--hostname', 'gitlab.com'],
-      { cwd: '/repo' }
-    )
-  })
-
-  it('enables creation for clean, in-sync, token-configured Azure DevOps feature branches', async () => {
-    mockAzureDevOpsProvider()
-
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'feature/azure',
-        base: 'main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'azure-devops',
-      canCreate: true,
-      blockedReason: null,
-      nextAction: null,
-      head: 'feature/azure'
-    })
-    expect(isAzureDevOpsReviewCreationAuthenticatedMock).toHaveBeenCalledOnce()
-    expect(ghExecFileAsyncMock).not.toHaveBeenCalled()
-    expect(glabExecFileAsyncMock).not.toHaveBeenCalled()
-  })
-
-  it('enables creation for clean, in-sync, token-configured Gitea feature branches', async () => {
-    mockGiteaProvider()
-
-    await expect(
-      getHostedReviewCreationEligibility({
-        repoPath: '/repo',
-        branch: 'feature/gitea',
-        base: 'main',
-        hasUncommittedChanges: false,
-        hasUpstream: true,
-        ahead: 0,
-        behind: 0
-      })
-    ).resolves.toMatchObject({
-      provider: 'gitea',
-      canCreate: true,
-      blockedReason: null,
-      nextAction: null,
-      head: 'feature/gitea'
-    })
-    expect(isGiteaReviewCreationAuthenticatedMock).toHaveBeenCalledOnce()
-    expect(ghExecFileAsyncMock).not.toHaveBeenCalled()
-    expect(glabExecFileAsyncMock).not.toHaveBeenCalled()
   })
 })
